2024/Stud/quals/WEB/SMM Auth/smm_service/requirements.txt
fastapi==0.105.0
SQLAlchemy==2.0.23
pydantic-settings==2.2.1
requests==2.31.0
uvicorn==0.29.0
Jinja2==3.1.4
python-multipart==0.0.9
python-jose==3.3.0